requirements.txt
numpy==1.23.3
Pillow==9.2.0
PyYAML==6.0
